+++ static/index.css
@@ -954,7 +954,7 @@ body {
        transition: all 0.7s ease;
 }
 .pbComment:hover .pbCommentContent{
-       border-radius: 0% 0% 50% 50% / 0% 0% 0% 5%;
+       border-radius: 0% 0% 50% 50% / 0% 0% 0% 0%;
        box-shadow: 10px 10px rgba(0, 0, 0, 0.25);
 }
 .pbComment:hover .pbCommentBotton{
