+++ static/js/blamer.js
@@ -345,121 +345,4 @@ export default class extends abstract {
             }
         }
     }
-    async findBlameThing(id, reactionData) {
-            console.log(reactionData.likeStatus)
-            let blameThing = "";
-            let comments = 0;
-            if (reactionData.comments) {
-                comments = reactionData.comments.length;
-            }
-
-            this.data.Posts.forEach((post) => {
-                if (post.PostId == id) {
-                    blameThing = `
-                            <div class="bPost">
-                            <div id="${id}" class="blameContent">
-                                   <div class="pbTop">
-                                          <div class="pbTitle">${post.Title}</div>
-                                          <div class="pbUsername">${post.Username}</div>
-                                          <div class="pbTime">${post.CreationTime}</div>
-                                   </div>
-                                   <textarea class="pbContent" readonly>${post.Content}</textarea>
-                                   <div class="pbBottom">
-                                          <div class="pbTopic">${post.Topics}</div>
-                                          <div class="pbLikeNumb"><span>${reactionData["likes"]} </span><span id="pbLikebtn">LIKE</span></div>
-                                          <div class="pbDislikeNumb"><span>${reactionData["dislikes"]} </span><span id="pbDislikebtn-">DISLIKE</span></div>
-                                          <div class="pbCommentNumb"><span>${comments} </span><span>COMMENTS</span></div>
-                                   </div>
-                                   </div>
-                                   </div>
-                            `;
-                }
-            });
-            return blameThing;
-        }
-        // eventlisterner for blame button = id: letsComment for sent comment
-    async createCommentBox(id) {
-        let container = document.createElement("div");
-        container.id = "blameC-form";
-        container.innerHTML += `
-              <div class="bPost" >
-              <div class="pbCommentBox">
-              <label id="letsComment" for="bCommentBoxContent" class="sendComment">Comment</label>
-              <input type="text" name="blameC-Content" class="pbCommentBoxContent" id="bCommentBoxContent" placeholder="Comment here"></input>
-              </div>
-              </div>
-              `
-        return container;
-    }
-
-    // search for comments
-    async findComments(commentData) {
-        let comments = "";
-        let side = "justify-items: start;";
-        let like = "white"
-        let dislike = "white"
-
-        commentData.forEach((comment) => {
-
-            if (comment.Username == this.activeUserName) {
-                side = "justify-items: end;";
-            }
-            if (comment.LikesStatus == "like") {
-                like = "green"
-            } else if (comment.LikesStatus == "dislike") {
-                dislike = "red"
-            }
-            comments += `
-                                   <div class="pbComment" style="${side};">
-                                          <div class="pbCommentUname"><b>${comment.Username}</b>${comment.CreationTime}</div>
-                                          <div class="pbCommentContent" >
-                                                 <div class="pbCommentText">${comment.Content}</div>
-                                          </div>
-                                          <div class="pbCommentBotton">
-                                                 <div class="pbCommentLike"><span id="lNumb">${comment.Likes}</span><span id="lButton" style="color :${like}">Like</span></div>
-                                                 <div class="pbCommentDislike"><span id="dNumb">${comment.Dislikes}</span><span id="lButton" style="color :${dislike}">Dislike</span></div>
-                                          </div>
-                                   </div>
-                            `;
-
-
-        });
-        return comments;
-    }
-    async createCommentArea(id, commentData) {
-        console.log("looking for commentID", id, commentData)
-
-        let comments = await this.findComments(commentData);
-        let parent = document.createElement("div");
-        parent.className = "bPost";
-        let container = document.createElement("div");
-        container.className = "pbCommentArea";
-        container.innerHTML = `${comments}`;
-        parent.appendChild(container);
-        return parent;
-    }
-    async blameContent(element, reactionData) {
-
-        console.log("reactionData", reactionData)
-        let parents = document.querySelectorAll(".bPost");
-        if (parents) {
-            parents.forEach((parent) => {
-                parent.remove();
-            });
-        }
-        let parent = document.getElementById("mainPostsBox");
-        let blame = await this.findBlameThing(element.id, reactionData);
-        parent.innerHTML = blame;
-        if (
-            this.activeUserName != null &&
-            this.activeUserName !== "guest"
-        ) {
-            console.log("in blameContent", this.activeUserName);
-            let commentBox = await this.createCommentBox(element.id);
-            parent.appendChild(commentBox);
-            if (reactionData.comments != null) {
-                parent.appendChild(await this.createCommentArea(element.id, reactionData.comments));
-            }
-        }
-    }
 }
